"quick constructors" + unified example
Instead of a `new` function which you have to pass a "mode" into,
"quick constructors" are a set of functions that set the underlying
`console::Term`. It's a little less code the user has to write and it's
got a nice "feel" to it.

All examples are now in a single example.rs file which also has an
equivalent `console::Term` example to show how `tiny-console` actually
provides a useful, simplified API over `console`.